sdk_python/pyproject.toml
[project]
name = "grctl-sdk-python"
version = "0.1.0"
description = "The Python SDK for the grctl platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.7.0",
    "msgspec>=0.18.0",
    "python-ulid>=2.0.0",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.15.9",
]


[tool.uv]
package = true

[tool.setuptools]
packages = ["grctl"]

[tool.pytest.ini_options]
addopts = "--ignore=tests/examples --ignore=tests/e2e"
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"
pythonpath = "."
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
]


[tool.ruff]
line-length = 120

lint.select = ["ALL"] # Include all the rules, including new ones
lint.ignore = [
    "BLE001",
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed
    "COM812",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D107",
    "D203",
    "D211",
    "D213",
    "EM101",
    "EM102",
    "G004",
    "TRY003",
    "TC001",
    "TC003",
    "S324",

]
[tool.ruff.lint.extend-per-file-ignores]
# Specific rules to ignore for tests
"tests/**/*.py" = [
    "S101",   # Allow assert in tests
    "ANN001", # Allow missing type annotations in tests
    "ANN201", # Allow missing type annotations in tests
    "ANN202", # Allow missing return type annotations in tests
    "ANN204", # Allow missing type annotations for self and cls in tests
    "ARG",    # Unused function args for fixtures
    "COM812",
    "FBT",    # Don't care about booleans as positional arguments in tests for @pytest.mark.parametrize
    "INP001", # Ignore that its not in a real python module. Doesn't matter for tests
    # The below are debateable
    "PLR2004", # Magic value used in comparison, ...
    "S311",    # Standard pseudo-random generators are not suitable for cryptographic purposes
    "ERA001",  # Commented out code
    "F841",    # Unused variable
    "PGH003",
    "SLF001", # Private member accessed outside of class
    "TRY002", # Create your own exception classes
]

"grctl/models/**/*.py" = [
    "ERA001", # Commented out code
]
